+++ spec/spec_helper.rb
@@ -5,6 +5,7 @@ SimpleCov.start do
   add_filter 'spec/'
   add_filter 'matchers.rb'
   add_filter '.direnv/'
+  add_filter 'libraries/resource_providers/'
   add_group 'Libraries', 'libraries'
   add_group 'Resources', 'resources'
   minimum_coverage 92 # Goal: A bit higher