+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Unreleased
   - Support oneview-sdk v3.0 & different API versions/modules
   - ethernet_network support for API300
+  - fc_network support for API300
+  - fcoe_network support for API300
+  - network_set support for API300
 
 ### 1.1.0
   - Add support for client ENV variables